Move mock_connection to cursor file
Instead of wondering where these things are located, move the mock into 
the file where its used
