+++ tests/conftest.py
@@ -194,16 +194,3 @@ def mock_transport(mocker):
 def mock_client(mocker):
     """A magicmock for pymapd.connection.Client"""
     return mocker.patch("pymapd.connection.Client")
-
-
-@pytest.fixture
-def mock_connection(mock_transport, mock_client):
-    """Connection with mocked transport layer, and
-
-    - username='user'
-    - password='password'
-    - host='localhost'
-    - dbname='dbname'
-    """
-    return connect(user='user', password='password',
-                   host='localhost', dbname='dbname')
